+++ src/resources/views/admin/modalities/create.blade.php
@@ -1,6 +1,6 @@
 @extends('layouts.admin')
 
-@section('titulo', 'Nova categoria')
+@section('titulo', 'Nova modalidade')
 @section('icone', 'flag')
 @section('subtitulo', $event->title)
 
@@ -10,14 +10,14 @@
     <div class="row justify-content-center">
         <div class="col-xl-8">
             <div class="card mb-4">
-                <div class="card-header">Dados da categoria</div>
+                <div class="card-header">Dados da modalidade</div>
                 <div class="card-body">
-                    <form method="POST" action="{{ route('admin.eventos.categorias.store', $event->id) }}">
+                    <form method="POST" action="{{ route('admin.eventos.modalidades.store', $event->id) }}">
                         @csrf
                         @include('admin.modalities._form')
                         <hr class="my-4">
-                        <button class="btn btn-primary" type="submit">Criar categoria</button>
-                        <a class="btn btn-link" href="{{ route('admin.eventos.categorias.index', $event->id) }}">Cancelar</a>
+                        <button class="btn btn-primary" type="submit">Criar modalidade</button>
+                        <a class="btn btn-link" href="{{ route('admin.eventos.modalidades.index', $event->id) }}">Cancelar</a>
                     </form>
                 </div>
             </div>